requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.10
djangorestframework==3.14.0
filelock==3.13.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==23.2
pillow==10.2.0
psycopg==3.1.18
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sqlparse==0.4.4
sympy==1.12
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.2
transformers==4.37.2
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
whitenoise==6.6.0
